Split test into logical parts
